Sale order: move more services to java packages
RM84578
