+++ axelor-sale/src/main/java/com/axelor/apps/sale/module/SaleModule.java
@@ -54,8 +54,6 @@ import com.axelor.apps.sale.service.ProductCategorySaleService;
 import com.axelor.apps.sale.service.ProductCategoryServiceSaleImpl;
 import com.axelor.apps.sale.service.ProductRestService;
 import com.axelor.apps.sale.service.ProductRestServiceImpl;
-import com.axelor.apps.sale.service.SaleOrderRestService;
-import com.axelor.apps.sale.service.SaleOrderRestServiceImpl;
 import com.axelor.apps.sale.service.app.AppSaleService;
 import com.axelor.apps.sale.service.app.AppSaleServiceImpl;
 import com.axelor.apps.sale.service.cart.CartCreateService;
@@ -115,20 +113,22 @@ import com.axelor.apps.sale.service.saleorder.SaleOrderInitValueService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderInitValueServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderMarginService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderMarginServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingService;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingViewService;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingViewServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderOnChangeService;
-import com.axelor.apps.sale.service.saleorder.SaleOrderOnChangeServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderOnLineChangeService;
-import com.axelor.apps.sale.service.saleorder.SaleOrderOnLineChangeServiceImpl;
+import com.axelor.apps.sale.service.saleorder.SaleOrderRestService;
+import com.axelor.apps.sale.service.saleorder.SaleOrderRestServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderUserService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderUserServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderVersionService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderVersionServiceImpl;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingService;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingServiceImpl;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingViewService;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingViewServiceImpl;
+import com.axelor.apps.sale.service.saleorder.onchange.SaleOrderOnChangeService;
+import com.axelor.apps.sale.service.saleorder.onchange.SaleOrderOnChangeServiceImpl;
+import com.axelor.apps.sale.service.saleorder.onchange.SaleOrderOnLineChangeService;
+import com.axelor.apps.sale.service.saleorder.onchange.SaleOrderOnLineChangeServiceImpl;
 import com.axelor.apps.sale.service.saleorder.opportunity.OpportunitySaleOrderService;
 import com.axelor.apps.sale.service.saleorder.opportunity.OpportunitySaleOrderServiceImpl;
 import com.axelor.apps.sale.service.saleorder.opportunity.OpportunityServiceSaleImpl;
