+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/SaleOrderRestService.java
@@ -1,4 +1,4 @@
-package com.axelor.apps.sale.service;
+package com.axelor.apps.sale.service.saleorder;
 
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.sale.db.SaleOrder;
