+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/merge/SaleOrderMergingViewServiceImpl.java
@@ -16,11 +16,11 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.axelor.apps.sale.service.saleorder;
+package com.axelor.apps.sale.service.saleorder.merge;
 
 import com.axelor.apps.sale.db.SaleOrder;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingService.Checks;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingService.SaleOrderMergingResult;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingService.Checks;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingService.SaleOrderMergingResult;
 import com.axelor.auth.db.AuditableModel;
 import com.axelor.i18n.I18n;
 import com.axelor.meta.schema.actions.ActionView;
