+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/onchange/SaleOrderOnChangeService.java
@@ -1,4 +1,4 @@
-package com.axelor.apps.sale.service.saleorder;
+package com.axelor.apps.sale.service.saleorder.onchange;
 
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.sale.db.SaleOrder;
