+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/module/SupplychainModule.java
@@ -60,12 +60,12 @@ import com.axelor.apps.sale.service.saleorder.SaleOrderComputeServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderCreateServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderInitValueServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderLineTreeComputationServiceSupplychainImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderMergingViewServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderOnChangeServiceImpl;
-import com.axelor.apps.sale.service.saleorder.SaleOrderOnLineChangeServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderServiceImpl;
 import com.axelor.apps.sale.service.saleorder.SaleOrderVersionServiceImpl;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingServiceImpl;
+import com.axelor.apps.sale.service.saleorder.merge.SaleOrderMergingViewServiceImpl;
+import com.axelor.apps.sale.service.saleorder.onchange.SaleOrderOnChangeServiceImpl;
+import com.axelor.apps.sale.service.saleorder.onchange.SaleOrderOnLineChangeServiceImpl;
 import com.axelor.apps.sale.service.saleorder.opportunity.OpportunitySaleOrderServiceImpl;
 import com.axelor.apps.sale.service.saleorder.print.SaleOrderProductPrintingService;
 import com.axelor.apps.sale.service.saleorder.print.SaleOrderProductPrintingServiceImpl;
